+++ boot-data-neo4j/src/test/java/com/example/demo/ContainersConfig.java
@@ -1,0 +1,19 @@
+package com.example.demo;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.context.annotation.Bean;
+import org.testcontainers.neo4j.Neo4jContainer;
+
+import java.time.Duration;
+
+@TestConfiguration(proxyBeanMethods = false)
+public class ContainersConfig {
+
+    @Bean
+    @ServiceConnection
+    Neo4jContainer neo4jContainer() {
+        return new Neo4jContainer("neo4j:5")
+                .withStartupTimeout(Duration.ofMinutes(5));
+    }
+}
