+++ boot-data-neo4j/src/test/java/com/example/demo/PostRepositoryWithTestcontainersTest.java
@@ -4,17 +4,11 @@ import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.data.neo4j.test.autoconfigure.DataNeo4jTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Sort;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.Neo4jContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 import reactor.core.publisher.Flux;
 import reactor.test.StepVerifier;
 
@@ -25,26 +19,10 @@ import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@SpringBootTest
-//@DataNeo4jTest
-@Testcontainers
+@DataNeo4jTest
+@Import(ContainersConfig.class)
 @Slf4j
-// Testcontainers does not work well with per_class testinstance.
-// see: https://stackoverflow.com/questions/61357116/exception-mapped-port-can-only-be-obtained-after-the-container-is-started-when/61358336#61358336
-@TestInstance(TestInstance.Lifecycle.PER_METHOD)
-@ActiveProfiles("test")
 public class PostRepositoryWithTestcontainersTest {
-
-    @Container
-    static Neo4jContainer<?> neo4jContainer = new Neo4jContainer<>("neo4j:5")
-            .withStartupTimeout(Duration.ofMinutes(5));
-
-    @DynamicPropertySource
-    static void neo4jProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.neo4j.uri", neo4jContainer::getBoltUrl);
-        registry.add("spring.neo4j.authentication.username", () -> "neo4j");
-        registry.add("spring.neo4j.authentication.password", neo4jContainer::getAdminPassword);
-    }
 
     @Autowired
     private PostRepository posts;
