+++ boot-data-neo4j/src/test/java/com/example/demo/TestDemoApplication.java
@@ -1,25 +1,11 @@
 package com.example.demo;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
-import org.springframework.context.annotation.Bean;
-import org.testcontainers.neo4j.Neo4jContainer;
 
-import java.time.Duration;
-
-@TestConfiguration(proxyBeanMethods = false)
 public class TestDemoApplication {
 
-    @Bean
-    @ServiceConnection
-    Neo4jContainer neo4jContainer() {
-        return new Neo4jContainer("neo4j:5")
-                .withStartupTimeout(Duration.ofMinutes(5));
-    }
-
     public static void main(String[] args) {
-        SpringApplication.from(DemoApplication::main).with(TestDemoApplication.class).run(args);
+        SpringApplication.from(DemoApplication::main).with(ContainersConfig.class).run(args);
     }
 
 }
